edit all files
