+++ sivkov.filipp/T2/Delimiter.hpp
@@ -1,0 +1,18 @@
+#ifndef DELIMITER_HPP
+#define DELIMITER_HPP
+#include <iostream>
+
+struct DelimiterI
+{
+  char expected;
+};
+
+struct DelimiterStr
+{
+  const char * expected;
+};
+
+std::istream& operator>>(std::istream& in, DelimiterI&& exp);
+std::istream& operator>>(std::istream& in, DelimiterStr&& exp);
+
+#endif
